Add WIP code to ProcessApiRequestCommandHandler that will match the incoming Request to the first Event by the Event's zip code. If a match is found, the Request is created in the db and associated to the Event. If a matching Event is not found by zip code, we do not create a record in the Request table.
- regardless of whether we decide to service the request or not, we still publish ApiRequestProcessedNotification and continue to invoke GASA's API with the appropriate Acceptance value (true for we can service the request, false for we can't service the request)
- renamed some messages/handlers to reflect their functionality
- udpated unit tests to compile and pass
