+++ AllReadyApp/Web-App/AllReady/Areas/Admin/Features/Requests/ImportRequestsCommandHandler.cs
@@ -45,6 +45,5 @@ namespace AllReady.Areas.Admin.Features.Requests
 
             return errors;
         }
-
     }
 }
