--- AllReadyApp/Web-App/AllReady/Features/Requests/ApiRequestAddedNotification.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using MediatR;
-
-namespace AllReady.Features.Requests
-{
-    public class ApiRequestAddedNotification : IAsyncNotification
-    {
-        public Guid RequestId { get; set; }
-    }
-}
