+++ AllReadyApp/Web-App/AllReady/Features/Requests/ApiRequestProcessedNotification.cs
@@ -1,0 +1,9 @@
+﻿using MediatR;
+
+namespace AllReady.Features.Requests
+{
+    public class ApiRequestProcessedNotification : IAsyncNotification
+    {
+        public string ProviderRequestId { get; set; }
+    }
+}
